survey api make private

+++ controllers/surveyController.js
@@ -6,17 +6,17 @@ import Survey from "../models/surveyModel.js";
 
 //@desc Get all surveys
 //@route Get /api/v1/surveys
-//@access Public
+//@access Private
 const getAllSurveys = asyncHandler(async (req, res)=>{
-    const survey = await Survey.find();
+    const survey = await Survey.find({user_id: req.user.id});
     res.status(200).json(survey);
 });
 
 //@desc Get a survey
 //@route Get /api/v1/surveys/:id
-//@access Public
+//@access Private
 const getASurvey = asyncHandler(async (req, res)=>{
-    const fetchedSurvey = await Survey.findById(req.params.id);
+    const fetchedSurvey = await Survey.findById({_id:req.params.id, user_id: req.user.id});
     if(!fetchedSurvey) {
         res.status(404);
         throw new Error("Survey not found.");
@@ -26,7 +26,7 @@ const getASurvey = asyncHandler(async (req, res)=>{
 
 //@desc Save a survey
 //@route Post /api/v1/surveys
-//@access Public
+//@access Private
 const saveSurvey = asyncHandler(async (req, res)=>{
     const {title, desc, author} = req.body;
     if(!title || !desc || !author) {
@@ -36,7 +36,8 @@ const saveSurvey = asyncHandler(async (req, res)=>{
     const savedSurvey = await Survey.create({
         title, 
         desc, 
-        author
+        author,
+        user_id: req.user.id,
     })
     res.status(201).json(savedSurvey);
 });
@@ -44,16 +45,16 @@ const saveSurvey = asyncHandler(async (req, res)=>{
 
 //@desc Update a survey
 //@route Put /api/v1/surveys/:id
-//@access Public
+//@access Private
 const updateSurvey = asyncHandler(async (req, res)=>{
-    const fetchedSurvey = await Survey.findById(req.params.id);
+    const fetchedSurvey = await Survey.findById({_id: req.params.id, user_id: req.user.id});
     if(!fetchedSurvey) {
         res.status(404);
         throw new Error("Survey not found.");
     }
 
     const updatedSurvey = await Survey.findByIdAndUpdate(
-        req.params.id,
+        {_id:req.params.id, user_id: req.user.id},
         req.body,
         {new: true}
     );
@@ -62,15 +63,15 @@ const updateSurvey = asyncHandler(async (req, res)=>{
 
 //@desc Delete a surveys
 //@route Delete /api/v1/surveys/:id
-//@access Public
+//@access Private
 const deleteSurvey = asyncHandler(async (req, res)=>{
-    const fetchedSurvey = await Survey.findById(req.params.id);
+    const fetchedSurvey = await Survey.findById({_id: req.params.id, user_id: req.user.id});
     if(!fetchedSurvey) {
         res.status(404);
         throw new Error("Survey not found.");
     }
 
-    await Survey.findByIdAndDelete(req.params.id);
+    await Survey.findByIdAndDelete({_id: req.params.id, user_id: req.user.id});
 
     res.status(200).json(fetchedSurvey);
 });
